Rename and export NxTransferListHalf - RSC-887
