+++ lib/src/components/NxTransferList/types.ts
@@ -12,33 +12,6 @@ import DataItem from '../../util/DataItem';
 
 export type FilterFn<T extends string | number = string> = (d: DataItem<T>[]) => DataItem<T>[];
 
-type SelectionChangeHandler<T> = (checked: boolean, id: T) => void;
-
-export interface TransferListItemProps<T extends string | number = string> extends DataItem<T> {
-  checked: boolean;
-  onChange: SelectionChangeHandler<T>;
-  showReorderingButtons?: boolean | null;
-  isFilteredItem?: boolean | null;
-  onReorderItem?: (index: number, direction: 1 | -1) => void | null;
-  index: number;
-  listLength: number;
-}
-
-export interface TransferListHalfProps<T extends string | number = string> {
-  label: Exclude<ReactNode, null | undefined>;
-  filterValue: string;
-  onFilterChange: NxFilterInputProps['onChange'];
-  showMoveAll: boolean;
-  onMoveAll: (toMove: T[]) => void;
-  items: DataItem<T>[];
-  isSelected: boolean;
-  onItemChange: SelectionChangeHandler<T>;
-  footerContent: ReactNode;
-  filterFn?: ((filterStr: string, itemDisplayName: string) => boolean) | null;
-  onReorderItem?: (index: number, direction: -1 | 1) => void | null;
-  allowReordering?: boolean | null;
-}
-
 export interface BaseStatefulProps<T extends string | number = string>
   extends Omit<HTMLAttributes<HTMLDivElement>, 'children' | 'onChange'> {
   allItems: DataItem<T>[];
