+++ lib/src/components/NxTransferListHalf/NxTransferListHalf.scss
